+++ src/cli/project/create/index.js
@@ -4,23 +4,33 @@ const { cli, action } = require('core/cli');
 const { execute, run } = require('./create');
 
 cli(
-  'create-project [targetDir]',
+  'project:create [starterName]',
   'Create a project from forge starter projects',
   input => action(execute, run, input),
   {
+    alias: 'create-project',
     requirements: {
       forgeRelease: false,
       runningNode: false,
       rpcClient: false,
       wallet: false,
     },
-    options: [['-y, --yes', 'Use default value for all options']],
+    options: [
+      ['-y, --yes', 'Use default value for all options'],
+      ['--target <target>', 'Target directory, default is current directory'],
+      ['--starterDir <starterDir>', 'Local boilerplate directory'],
+      ['--registry <registry>', 'Only for npm package'],
+    ],
     handlers: {
       '--help': () => {
         shell.echo(`
 Examples:
-  - ${chalk.cyan('forge create-project')}           create a project interactively
-  - ${chalk.cyan('forge create-project --yes')}      create a project with default settings
+  - ${chalk.cyan(
+    'forge project:create forge-react-starter'
+  )}           create a project by forge-react-starter template interactively
+  - ${chalk.cyan(
+    'forge project:create forge-react-starter --yes'
+  )}      create a project by forge-react-starter template with default settings
         `);
       },
     },